requirements.txt
# Core
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic>=2.7.0

# LLM clients
openai>=1.0.0
boto3>=1.34.0

# Resume parsing
pdfplumber>=0.11.0
docx2txt>=0.8

# HTTP
requests>=2.31.0
httpx>=0.27.0

# STT fallback
elevenlabs>=1.0.0

# TTS (local, optional — install separately if needed)
# pocket-tts
# scipy
# numpy

# TTS (Mistral Voxtral — uses requests, no extra package)

# Database (optional — runs without DB if DATABASE_URL not set)
psycopg2-binary>=2.9.0
